backend/requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
python-multipart==0.0.6
psycopg2-binary==2.9.7
python-dotenv==1.0.0
